+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/general/HideLayoutComponentsPatch.kt
@@ -27,7 +27,14 @@ import com.android.tools.smali.dexlib2.iface.instruction.FiveRegisterInstruction
         LithoFilterPatch::class,
         SettingsPatch::class
     ],
-    compatiblePackages = [CompatiblePackage("com.google.android.youtube", ["18.32.39"])]
+    compatiblePackages = [
+        CompatiblePackage(
+            "com.google.android.youtube", [
+                "18.32.39",
+                "18.37.36"
+            ]
+        )
+    ]
 )
 @Suppress("unused")
 object HideLayoutComponentsPatch : BytecodePatch(
@@ -43,6 +50,40 @@ object HideLayoutComponentsPatch : BytecodePatch(
                 "revanced_hide_gray_separator_title",
                 "revanced_hide_gray_separator_summary_on",
                 "revanced_hide_gray_separator_summary_off",
+            ),
+            SwitchPreference(
+                "revanced_hide_join_membership_button",
+                StringResource("revanced_hide_join_membership_button_title", "Hide \"Join\" button"),
+                StringResource("revanced_hide_join_membership_button_summary_on", "Button is hidden"),
+                StringResource("revanced_hide_join_membership_button_summary_off", "Button is shown")
+            ),
+
+            SwitchPreference(
+                "revanced_hide_notify_me_button",
+                StringResource("revanced_hide_notify_me_button_title", "Hide \"Notify me\" button"),
+                StringResource("revanced_hide_notify_me_button_summary_on", "Button is hidden"),
+                StringResource("revanced_hide_notify_me_button_summary_off", "Button is shown")
+            ),
+            SwitchPreference(
+                "revanced_hide_timed_reactions",
+                StringResource("revanced_hide_timed_reactions_title", "Hide timed reactions"),
+                StringResource("revanced_hide_timed_reactions_summary_on", "Timed reactions are hidden"),
+                StringResource("revanced_hide_timed_reactions_summary_off", "Timed reactions are shown")
+            ),
+            SwitchPreference(
+                "revanced_hide_search_result_shelf_header",
+                StringResource(
+                    "revanced_hide_search_result_shelf_header_title",
+                    "Hide search result shelf header"
+                ),
+                StringResource(
+                    "revanced_hide_search_result_shelf_header_summary_on",
+                    "Shelf header is hidden"
+                ),
+                StringResource(
+                    "revanced_hide_search_result_shelf_header_summary_off",
+                    "Shelf header is shown"
+                )
             ),
             SwitchPreference(
                 "revanced_hide_channel_guidelines",