+++ src/main/kotlin/app/revanced/patches/youtube/layout/sponsorblock/SponsorBlockBytecodePatch.kt
@@ -44,7 +44,8 @@ import com.android.tools.smali.dexlib2.iface.reference.StringReference
                 "18.20.39",
                 "18.23.35",
                 "18.29.38",
-                "18.32.39"
+                "18.32.39",
+                "18.37.36"
             ]
         )
     ],
@@ -95,11 +96,8 @@ object SponsorBlockBytecodePatch : BytecodePatch(
 
         /*
          * Set current video id.
-         *
-         * The new video id hook seems to work without issues,
-         * but it's easier to keep using this hook as it's well tested and has no known problems.
          */
-        VideoIdPatch.legacyInjectCallBackgroundPlay("$INTEGRATIONS_SEGMENT_PLAYBACK_CONTROLLER_CLASS_DESCRIPTOR->setCurrentVideoId(Ljava/lang/String;)V")
+        VideoIdPatch.hookBackgroundPlayVideoId("$INTEGRATIONS_SEGMENT_PLAYBACK_CONTROLLER_CLASS_DESCRIPTOR->setCurrentVideoId(Ljava/lang/String;)V")
 
         /*
          * Seekbar drawing